+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
-v1.0.11hotfix2
+
+## v1.0.12beta
+- fixed unloadable settings for 64bit DE and FR clients (bugfix)
+- fixed label of Hultvis hunter guide in FR client (enhancement)
+
+## v1.0.11hotfix2
 - fixing plugincompendium version data
 
 ## v1.0.11hotfix1
@@ -7,8 +12,7 @@ v1.0.11hotfix2
 
 ## v1.0.11beta
 - full U29 travel skills support for DE and FR (enhancement)
-## v1.0.11beta
-- full U29 travel skills support for DE and FR (enhancement)
+
 ## v1.0.10beta
 - U29 travel skills support for EN (enhancement)
 - attempted U29 travel skills support for DE and FR - no translations for skills were available on Bullroarer (enhancement)